implemented button color change to make sure remove account, though going to change to just pressing remove button

+++ app/src/main/java/org/michiganhackers/michiganhackers/SettingsFragment.java
@@ -1,6 +1,8 @@
 package org.michiganhackers.michiganhackers;
 
 import android.content.Intent;
+import android.graphics.Color;
+import android.graphics.PorterDuff;
 import android.os.Bundle;
 import android.support.annotation.NonNull;
 import android.support.v7.app.AppCompatActivity;
@@ -113,6 +115,9 @@ public class SettingsFragment extends Fragment {
                 changePassword.setVisibility(View.GONE);
                 sendEmail.setVisibility(View.GONE);
                 remove.setVisibility(View.GONE);
+
+                btnRemoveUser.getBackground().clearColorFilter();
+                btnRemoveUser.setText(R.string.remove_user);
             }
         });
 
@@ -153,6 +158,9 @@ public class SettingsFragment extends Fragment {
                 changePassword.setVisibility(View.VISIBLE);
                 sendEmail.setVisibility(View.GONE);
                 remove.setVisibility(View.GONE);
+
+                btnRemoveUser.getBackground().clearColorFilter();
+                btnRemoveUser.setText(R.string.remove_user);
             }
         });
 
@@ -210,6 +218,9 @@ public class SettingsFragment extends Fragment {
                 changePassword.setVisibility(View.GONE);
                 sendEmail.setVisibility(View.VISIBLE);
                 remove.setVisibility(View.GONE);
+
+                btnRemoveUser.getBackground().clearColorFilter();
+                btnRemoveUser.setText(R.string.remove_user);
             }
         });
 
@@ -242,7 +253,7 @@ public class SettingsFragment extends Fragment {
             @Override
             public void onClick(View v) {
                 progressBar.setVisibility(View.VISIBLE);
-                if (user != null) {
+                if (user != null && btnRemoveUser.getText().toString().equals(getString(R.string.confirm_remove_user))) {
                     user.delete()
                             .addOnCompleteListener(new OnCompleteListener<Void>() {
                                 @Override
@@ -259,6 +270,9 @@ public class SettingsFragment extends Fragment {
                                 }
                             });
                 }
+                btnRemoveUser.setText(R.string.confirm_remove_user);
+                btnRemoveUser.getBackground().setColorFilter(getResources().getColor(R.color.btn_remove_user_confirm), PorterDuff.Mode.MULTIPLY);
+                progressBar.setVisibility(View.GONE);
             }
         });
 
@@ -268,7 +282,6 @@ public class SettingsFragment extends Fragment {
                 signOut();
             }
         });
-
 
         return layout;
     }
